Fix database port type for PHP 8.3 MySQL connection

+++ app/Config/Database.php
@@ -207,7 +207,7 @@ class Database extends Config
             $this->default['username'] = getenv('MYSQLUSER') ?: getenv('MYSQL_USER') ?: '';
             $this->default['password'] = getenv('MYSQLPASSWORD') ?: getenv('MYSQL_PASSWORD') ?: '';
             $this->default['database'] = getenv('MYSQLDATABASE') ?: getenv('MYSQL_DATABASE') ?: '';
-            $this->default['port']     = getenv('MYSQLPORT') ?: getenv('MYSQL_PORT') ?: 3306;
+            $this->default['port']     = (int) (getenv('MYSQLPORT') ?: getenv('MYSQL_PORT') ?: 3306);
         }
     }
 }
